docs(icon-demo): add an icon demo to display all icons
