+++ apps/knapsack/data/blocks/block.FvKIrSQgq0.json
@@ -1,0 +1,10 @@
+{
+  "id": "FvKIrSQgq0",
+  "blockType": "pattern-template-block",
+  "data": {
+    "patternId": "icon",
+    "demoId": "iPghK4UYBt",
+    "templateId": "web-components-Pnw8Nf0oOd",
+    "hideControls": true
+  }
+}
