+++ apps/knapsack/data/demos/demo.iPghK4UYBt.json
@@ -1,0 +1,11 @@
+{
+  "id": "iPghK4UYBt",
+  "title": "Main",
+  "type": "data",
+  "patternId": "icon",
+  "templateId": "web-components-Pnw8Nf0oOd",
+  "data": {
+    "props": {},
+    "slots": {}
+  }
+}
