+++ libs/components/src/icon/icon-demo.scss
@@ -1,0 +1,25 @@
+:host {
+  --mdc-icon-font: 'covalent-icons';
+}
+
+.icon-grid {
+  display: flex;
+  flex-wrap: wrap; /* Enables wrapping to new rows */
+  gap: 24px; /* Space between items */
+  justify-content: start; /* Align items to the left */
+}
+
+.icon-container {
+  margin: 20px 5px;
+  text-align: center;
+  width: 200px;
+
+  cv-icon {
+    margin-bottom: 16px;
+    font-size: 24px;
+  }
+}
+
+.wrap {
+  word-wrap: break-word;
+}
