+++ libs/components/vite.config.js
@@ -31,6 +31,7 @@ module.exports = defineConfig(({ mode }) => {
           'libs/components/src/formfield/formfield',
           'libs/components/src/full-screen-dialog/full-screen-dialog',
           'libs/components/src/icon/icon',
+          'libs/components/src/icon/icon-demo',
           'libs/components/src/icon-button/icon-button',
           'libs/components/src/icon-button-toggle/icon-button-toggle',
           'libs/components/src/icon-checkbox/icon-check-toggle',
